Implemented services as DAL, this finishes the architecture
