+++ lib/models/example/exampleModel.ts
@@ -1,0 +1,14 @@
+import * as mongoose from "mongoose";
+
+const Schema = mongoose.Schema;
+
+export const ExampleSchema = new Schema({
+    element: {
+        type: String,
+        required: "Enter element1"
+    },
+    note: {
+        type: String,
+        required: "Enter testVar2"
+    }
+});
