--- lib/routes/exampleRoutes.ts
@@ -1,12 +1,0 @@
-import * as testController from "../controllers/exampleController";
-
-export class Routes {
-    private ctrlFunc = new testController.TestController();
-
-    public routes(app) {
-        // Hello World example route, with controller
-        app.route("/")
-            .get(this.ctrlFunc.testRouteFunction)
-            .post(this.ctrlFunc.testRouteFunction);
-    }
-}
